[IMP] mail: move parseEmail function
This commit moves parseEmail() from deprecated file `@mail/js/utils.js` to `@mail/utils/common/format.js`
